Basic Movement
Added PacmanController class. Pac-Man can move around the screen and once a direction is set, Pac-Man continues in that direction. Future changes will stop this movement when hitting a wall or if a new input would result in Pac-Man hitting a wall.

Created a MazeTile class. This will store data about every tile in the maze but has not been implemented yet.