Tile origin same as maze origin (top left)
Tiles now start at the top left of the maze. This should scale with the position/scale of the maze and tiles.
The scale of tiles is still 1x1, so this needs to be adjusted and also scaled to the size of the maze.
MazeGenerator has an explanation of i and j being swapped in coordinates and not when searching array. On a related note, I learned something new about the order in which two-dimensional arrays are searched today...